c-bitmap: round up bit-count for *_all() helpers
If you call c_bitmap_{set,clear}_all() with a bit-count that is not a
multiple of 8, we still know that you must have allocated enough bytes to
hold the specified bits. We define bitmaps as an array of bytes, so
alignment is guaranteed.

Make sure we always round up the bit-count to the next multiple of 8 so we
explicitly modify all requested bits plus the padding, rather than
skipping the high bits.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>