c-list: add c_list_entry_is_linked()
Mark entries as unlinked by pointing them to themselves, allowing us
to distinguish them from entries that are linked as the only entry
on a list (hence having both prev end next set to NULL).

Use this in c_list_remove() to not attempt to remove an entry that is
not linked on any list.

Signed-off-by: Tom Gundersen <teg@jklm.no>