c-macro: avoid __typeof__ in c_container_of()
The __typeof__ helper strips some qualifiers, including _Atomic [1].
This is useful in most scenarios, but not in c_container_of(), where we
want to explicitly check that the given pointer has a matching type.
Unlike `const' et.al. the _Atomic qualifier is not upgradable, so you
get warnings if the types don't match.

Avoid using __typeof__ and instead use __auto_type with a dead-code
expression to verify the types match.

[1] See gcc:
    * cp/parser.c (cp_parser_simple_type_specifier): Change TYPEOF for
      atomic types to return the non-atomic type.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>