c-string: beef up utf8 tests
Test (almost) every 1,2,3,4-byte sequence to verify that they are correctly classified
as either valid or invalid UTF-8.

For the sake of simplicity we don't test 4-bytes sequences that are prefixed
with a valid 3-byte sequence, etc.

For the sake of speed we only check every 64th 4-byte sequence by default.

Signed-off-by: Tom Gundersen <teg@jklm.no>