test: work around GCC bug in __builtin_constant_p()
The __builtin_constant_p() expression checks whether its argument is
a constant expression or not. The result of the builtin is always a
constant expression, since the fact whether its expression is constant
or not, is itself constant. This is documented in the formal definition
of the builtin, see:

    https://gcc.gnu.org/ml/gcc-patches/2008-10/msg01061.html

However, unfortunately, gcc is not very good in adhering to that.
Whenever optimizations are enabled, those invariants are no longer
true. This is a known gcc bug:

    https://gcc.gnu.org/bugzilla/show_bug.cgi?id=38377

Lets work around it and stop using __builtin_constant_p() in constant
expressions.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>